TriangulationInfoCache -> GridTools::Cache
