+++ source/grid/grid_tools_cache.inst.in
@@ -18,6 +18,6 @@
 for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
 {
 #if deal_II_dimension <= deal_II_space_dimension
-    template class TriangulationInfoCache<deal_II_dimension, deal_II_space_dimension>;
+    template class Cache<deal_II_dimension, deal_II_space_dimension>;
 #endif
 }
